requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
Flask==3.1.0
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
openai==1.69.0
pydantic==2.11.1
pydantic_core==2.33.0
pyee==11.1.1
python-dotenv==1.1.0
requests==2.32.3
setuptools==78.1.0
slack_bolt==1.23.0
slack_sdk==3.35.0
slackeventsapi==3.0.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
